requirements.txt
aiohttp
python-dotenv
PyMuPDF
python-docx
fastapi
uvicorn
pydantic
typing_extensions
regex
rich
aiofiles
requests
loguru